backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
openai>=1.12.0
groq>=0.4.0
python-dotenv==1.0.0
python-multipart==0.0.20
yfinance>=0.2.40
certifi>=2023.7.22
requests>=2.31.0
urllib3>=2.0.0
lxml>=4.9.0
pandas>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-search-results>=2.4.2
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
# Audio preprocessing dependencies removed - not needed for basic transcription
